# Restructure the class Geopotential